fix moved files
